// projects/jb-component-library/src/lib/jb-utils/keyboard-key.enum.ts
// See full list of keys here: https://developer.mozilla.org/en-US/docs/Web/API/KeyboardEvent/key/Key_Values

// When doing switch cases, make sure to include the older browser variants
export enum KeyboardKey {
  ArrowLeft = 'ArrowLeft',
  ArrowLeftIE = 'Left',
  ArrowDown = 'ArrowDown',
  ArrowDownIE = 'Down',
  ArrowRight = 'ArrowRight',
  ArrowRightIE = 'Right',
  ArrowUp = 'ArrowUp',
  ArrowUpIE = 'Up',
  Backspace = 'Backspace',
  Enter = 'Enter',
  Escape = 'Escape',
  EscapeIE = 'Esc',
  End = 'End',
  Home = 'Home',
  Shift = 'Shift',
  Space = ' ',
  SpaceIE = 'SpaceBar',
  Tab = 'Tab',
}